# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LicenseDuration = typeof LicenseDuration[keyof typeof LicenseDuration];


export const LicenseDuration = {
  '1month': '1month',
  lifetime: 'lifetime',
} as const;

export type LicenseStatus = typeof LicenseStatus[keyof typeof LicenseStatus];


export const LicenseStatus = {
  unused: 'unused',
  active: 'active',
  revoked: 'revoked',
  expired: 'expired',
  not_found: 'not_found',
} as const;

export interface LicenseGenerateRequest {
  /**
     * @minimum 1
     * @maximum 100
     */
  count?: number;
  duration: LicenseDuration;
  adminSecret: string;
}

export interface LicenseGenerateResponse {
  keys: string[];
}

export interface LicenseActivateRequest {
  key: string;
  deviceFingerprint?: string;
}

export interface LicenseActivationResponse {
  valid: boolean;
  duration?: LicenseDuration;
  activatedAt?: string;
  expiresAt?: string;
}

export interface LicenseValidateRequest {
  key: string;
}

export interface LicenseValidationResponse {
  valid: boolean;
  status?: LicenseStatus;
  duration?: LicenseDuration;
  activatedAt?: string;
  expiresAt?: string;
}