# lib/api-zod/src/generated/types/licenseActivationResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LicenseDuration } from './licenseDuration';

export interface LicenseActivationResponse {
  valid: boolean;
  duration?: LicenseDuration;
  activatedAt?: Date;
  expiresAt?: Date;
}