# lib/api-zod/src/generated/types/licenseDuration.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type LicenseDuration = typeof LicenseDuration[keyof typeof LicenseDuration];


export const LicenseDuration = {
  '1month': '1month',
  lifetime: 'lifetime',
} as const;