# lib/api-zod/src/generated/types/licenseGenerateRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LicenseDuration } from './licenseDuration';

export interface LicenseGenerateRequest {
  /**
     * @minimum 1
     * @maximum 100
     */
  count?: number;
  duration: LicenseDuration;
  adminSecret: string;
}