# lib/api-zod/src/generated/types/licenseStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type LicenseStatus = typeof LicenseStatus[keyof typeof LicenseStatus];


export const LicenseStatus = {
  unused: 'unused',
  active: 'active',
  revoked: 'revoked',
  expired: 'expired',
  not_found: 'not_found',
} as const;