# lib/api-zod/src/generated/types/licenseValidationResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LicenseDuration } from './licenseDuration';
import type { LicenseStatus } from './licenseStatus';

export interface LicenseValidationResponse {
  valid: boolean;
  status?: LicenseStatus;
  duration?: LicenseDuration;
  activatedAt?: Date;
  expiresAt?: Date;
}